move boxsnapguides to seperate component to prevent box rendering style flickers remove box border changing on snap to fix the same render err
